requirements.txt
qwen-agent
python-dotenv